Added generic type 2 appimage integration Updated squashfs library for increased performance Slight tweaking of Icon() to more often get the correct icon.
